+++ gears/carto_gears_api/lib/carto_gears_api/helpers/gear_controller_helper.rb
@@ -1,0 +1,24 @@
+module CartoGearsApi
+  module Helpers
+    # This should be included in every controller using the following CARTO layouts:
+    # - carto_gears_api/dashboard
+    # - carto_gears_api/profile
+    module GearControllerHelper
+      include SafeJsObject
+      include CartoDB::ConfigUtils
+      include TrackjsHelper
+      include GoogleAnalyticsHelper
+      include HubspotHelper
+      include FrontendConfigHelper
+      include AppAssetsHelper
+      include MapsApiHelper
+      include SqlApiHelper
+      include CartoGearsApi::Helpers::UrlHelper
+
+      # @return [CartoGearsApi::Users::User] Logged user, `nil` if none.
+      def logged_user
+        @logged_user ||= CartoGearsApi::UsersService.new.logged_user(request)
+      end
+    end
+  end
+end
