+++ gears/carto_gears_api/lib/carto_gears_api/organizations/organizations.rb
@@ -1,0 +1,12 @@
+require 'values'
+require 'uuidtools'
+
+module CartoGearsApi
+  module Organizations
+    # Organization information.
+    #
+    # @attr_reader [String] name Organization name.
+    class Organization < Value.new(:name);
+    end
+  end
+end
